Fix version checks in BlenderMesh plugin
